Moved Vec3GLf struct to the Utils class; Added some useful operators to the Vec3GLf struct; General code cleanup

+++ SnowEngine/SnowEngine/World.h
@@ -7,7 +7,7 @@
 class World
 {
 private:
-	double _world_gravity;
+	GLfloat _world_gravity;
 	std::vector<Entity*> _entity_list;
 public:
 	World(double gravity);
